--- a/api/client_test.go
+++ b/api/client_test.go
@@ -20,8 +20,6 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"testing"
-
-	"github.com/prometheus/tsdb/testutil"
 )
 
 func TestConfig(t *testing.T) {
@@ -148,7 +146,9 @@ func TestDoGetFallback(t *testing.T) {
 	defer server.Close()
 
 	u, err := url.Parse(server.URL)
-	testutil.Ok(t, err)
+	if err != nil {
+		t.Fatal(err)
+	}
 	client := &httpClient{client: *(server.Client())}
 
 	// Do a post, and ensure that the post succeeds.
@@ -158,7 +158,7 @@ func TestDoGetFallback(t *testing.T) {
 	}
 	resp := &testResponse{}
 	if err := json.Unmarshal(b, resp); err != nil {
-		testutil.Ok(t, err)
+		t.Fatal(err)
 	}
 	if resp.Method != http.MethodPost {
 		t.Fatalf("Mismatch method")
@@ -174,7 +174,7 @@ func TestDoGetFallback(t *testing.T) {
 		t.Fatalf("Error doing local request: %v", err)
 	}
 	if err := json.Unmarshal(b, resp); err != nil {
-		testutil.Ok(t, err)
+		t.Fatal(err)
 	}
 	if resp.Method != http.MethodGet {
 		t.Fatalf("Mismatch method")