--- a/go.mod
+++ b/go.mod
@@ -4,8 +4,12 @@ require (
 	github.com/beorn7/perks v1.0.0
 	github.com/go-logfmt/logfmt v0.4.0 // indirect
 	github.com/golang/protobuf v1.3.1
+	github.com/json-iterator/go v1.1.6
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.1 // indirect
 	github.com/prometheus/client_model v0.0.0-20190129233127-fd36f4220a90
-	github.com/prometheus/common v0.4.0
-	github.com/prometheus/procfs v0.0.0-20190507164030-5867b95ac084
+	github.com/prometheus/common v0.4.1
+	github.com/prometheus/procfs v0.0.0-20190522114515-bc1a522cf7b1
 	github.com/prometheus/tsdb v0.7.1
+	github.com/stretchr/testify v1.3.0 // indirect
 )