Signed-off-by: Thomas Jackson <jacksontj.89@gmail.com>

Fixes #565