@krasi-georgiev 2019-05-08
do you think you can add some benchmarking so we can see what is the actual performance gains?