@jacksontj 2019-05-08
sure, and for easy reading here are the result:

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/client_golang/api/prometheus/v1
BenchmarkSerialization/10/10/unmarshal/encoding/json-4         	  200000	     10272 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/10/10/unmarshal/jsoniter-4              	10000000	       154 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/10/100/unmarshal/encoding/json-4        	   20000	     87834 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/10/100/unmarshal/jsoniter-4             	10000000	       178 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/10/1000/unmarshal/encoding/json-4       	    2000	    966855 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/10/1000/unmarshal/jsoniter-4            	 5000000	       211 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/100/10/unmarshal/encoding/json-4        	   10000	    124487 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/100/10/unmarshal/jsoniter-4             	10000000	       217 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/100/100/unmarshal/encoding/json-4       	    1000	   1335330 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/100/100/unmarshal/jsoniter-4            	10000000	       184 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/100/1000/unmarshal/encoding/json-4      	     100	  11345940 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/100/1000/unmarshal/jsoniter-4           	 5000000	       219 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/1000/10/unmarshal/encoding/json-4       	    1000	   1093244 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/1000/10/unmarshal/jsoniter-4            	10000000	       170 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/1000/100/unmarshal/encoding/json-4      	     200	   9597895 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/1000/100/unmarshal/jsoniter-4           	10000000	       180 ns/op	      16 B/op	       1 allocs/op
BenchmarkSerialization/1000/1000/unmarshal/encoding/json-4     	      10	 104403258 ns/op	     232 B/op	       5 allocs/op
BenchmarkSerialization/1000/1000/unmarshal/jsoniter-4          	10000000	       176 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/prometheus/client_golang/api/prometheus/v1	35.976s
```