@jacksontj 2019-05-09
Fixed bug in benchmarks, for some reason the difference isn't as large as I expected:

```
$ go test -run=x -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/prometheus/client_golang/api/prometheus/v1
BenchmarkSerialization/10/10/unmarshal/encoding/json-8         	    5000	    258376 ns/op	   25970 B/op	     564 allocs/op
BenchmarkSerialization/10/10/unmarshal/jsoniter-8              	   10000	    231181 ns/op	   32649 B/op	     850 allocs/op
BenchmarkSerialization/10/100/unmarshal/encoding/json-8        	     500	   2320385 ns/op	  256520 B/op	    5964 allocs/op
BenchmarkSerialization/10/100/unmarshal/jsoniter-8             	    1000	   2187491 ns/op	  327974 B/op	    8950 allocs/op
BenchmarkSerialization/10/1000/unmarshal/encoding/json-8       	      50	  23265065 ns/op	 2571482 B/op	   59970 allocs/op
BenchmarkSerialization/10/1000/unmarshal/jsoniter-8            	      50	  22021770 ns/op	 3287473 B/op	   89955 allocs/op
BenchmarkSerialization/100/10/unmarshal/encoding/json-8        	     500	   2568609 ns/op	  258018 B/op	    5606 allocs/op
BenchmarkSerialization/100/10/unmarshal/jsoniter-8             	    1000	   2239641 ns/op	  326580 B/op	    8501 allocs/op
BenchmarkSerialization/100/100/unmarshal/encoding/json-8       	      50	  23747338 ns/op	 2576962 B/op	   59650 allocs/op
BenchmarkSerialization/100/100/unmarshal/jsoniter-8            	      50	  22705622 ns/op	 3288373 B/op	   89526 allocs/op
BenchmarkSerialization/100/1000/unmarshal/encoding/json-8      	       5	 233949527 ns/op	26711548 B/op	  600269 allocs/op
BenchmarkSerialization/100/1000/unmarshal/jsoniter-8           	       5	 228011137 ns/op	33462769 B/op	  899805 allocs/op
BenchmarkSerialization/1000/10/unmarshal/encoding/json-8       	      50	  26338162 ns/op	 2596940 B/op	   56224 allocs/op
BenchmarkSerialization/1000/10/unmarshal/jsoniter-8            	      50	  23486559 ns/op	 3282834 B/op	   85182 allocs/op
BenchmarkSerialization/1000/100/unmarshal/encoding/json-8      	       5	 241223130 ns/op	27157049 B/op	  678011 allocs/op
BenchmarkSerialization/1000/100/unmarshal/jsoniter-8           	       5	 227933347 ns/op	33707764 B/op	  974806 allocs/op
BenchmarkSerialization/1000/1000/unmarshal/encoding/json-8     	       1	2324619810 ns/op	311494832 B/op	 6319035 allocs/op
BenchmarkSerialization/1000/1000/unmarshal/jsoniter-8          	       1	2155111994 ns/op	361141288 B/op	 9300019 allocs/op
PASS
ok  	github.com/prometheus/client_golang/api/prometheus/v1	34.515s
```


I'll need to spend some more time on this later :/