@jacksontj 2019-05-15
Some, at this point I can confirm that the benchmarks are accurate (meaning the perf improvement isn't what I expected). So this will require more investigation, but its possible changes will be required elsewhere as well (e.g. common package). I'll try to get some more time to investigate this week.