@jacksontj 2019-05-15
I added Type Encoder/Decoder funcs and am now seeing the perf improvement I was expecting. This seems like a pretty ugly way to do this (since we have to copy/paste this marshal logic all over). So maybe it would be helpful to change the marshaling/unmarshaling in common to be fast?

```
benchmark                                                 old ns/op      new ns/op     delta
BenchmarkSerialization/1000/1000/marshal/jsoniter-8       4125222362     201316013     -95.12%
BenchmarkSerialization/1000/1000/unmarshal/jsoniter-8     2060620113     218847915     -89.38%

benchmark                                                 old allocs     new allocs     delta
BenchmarkSerialization/1000/1000/marshal/jsoniter-8       19993541       2030           -99.99%
BenchmarkSerialization/1000/1000/unmarshal/jsoniter-8     9240110        998003         -89.20%

benchmark                                                 old bytes     new bytes     delta
BenchmarkSerialization/1000/1000/marshal/jsoniter-8       689480872     106189225     -84.60%
BenchmarkSerialization/1000/1000/unmarshal/jsoniter-8     361144008     9749326       -97.30%
```